Adapted OpenCL accelerator to properties/device model.
